requirements.txt
alembic==1.13.1
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
cornice==6.1.0
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.43.0
google-generativeai==0.3.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
hupper==1.12.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
PasteDeploy==3.1.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyramid==2.0.2
pyramid_tm==2.6
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9.1
SQLAlchemy==2.0.25
tqdm==4.67.1
transaction==5.0
translationstring==1.4
typing_extensions==4.15.0
urllib3==2.6.2
venusian==3.1.1
waitress==3.0.2
WebOb==1.8.9
zope.deprecation==6.0
zope.interface==8.1.1
zope.sqlalchemy==3.1
